Add base plane and lights
Add a base plane where all the objects will be placed.

Add a SpotLight to add a soft illumination to the plane and make it
visible.

Add a second spotlight which target the point where the ray casted from
the mouse position hits the plane.

Create a class to add such objects to the scene.

Minor changes and fixes.