Major code refactorings
Refactor many pieces of code extracting login from the script.js file to
split it into dedicated files.

Add a logging configuration file and a logging manager which reads it to
provide configurable logging functionalities.

Add a particle system manager to handle the creation and the lifetime of
the main particles mesh.

Add a scene manager (still unsure about its usefulness and unused as of
now).

Update experience.js to use all the auxiliary modules to render the
scene.

Make enough changes for the experience to at least run, many features
like the debug gui and the mouse interactability are still missing.

Minor changes and fixes.