Add animations
Add a file to manage animations for the main mesh and the title.

Use a promise to load the font and to signal that the title has been
added to the scene to allow other code to run when these operations have
completed.

Update code in experience.js to trigger animations and correctly
retrieve the title mesh instance once it has been loaded.

Minor changes and fixes.